+++ src/symtbl.h
@@ -1,13 +1,14 @@
 #pragma once
 
-#include "hashtbl.h"
+#include "../lib/hashtbl.h"
+#include "types.h"
 
 // Forward definition see parser.h
 typedef struct scope_t scope_t;
 
 typedef struct {
-  size_t line;
-  size_t column;
+  uint32_t line;
+  uint32_t column;
 } symbol_table_reference_t;
 
 // Basic symbol table entry.
@@ -25,13 +26,13 @@ typedef struct {
   str *name;
   e_ika_type type;
   void *storage;
-  size_t bytes; // NOTE: bits might be better in the long run
-  size_t alignment;
-  size_t line;
-  size_t column;
+  uint32_t bytes; // NOTE: bits might be better in the long run
+  uint32_t alignment;
+  uint32_t line;
+  uint32_t column;
   bool constant;
   symbol_table_reference_t **references;
-  size_t reference_count;
+  uint32_t reference_count;
 } symtbl_entry_t;
 
 typedef struct {
@@ -42,81 +43,11 @@ typedef struct {
 
 symtbl_t *symtbl_init(allocator_t allocator, scope_t *scope);
 
-void symtbl_insert(symtbl_t *, str name, e_ika_type type,
-                   bool constant, size_t line, size_t column);
+void symtbl_insert(symtbl_t *, str name, e_ika_type type, bool constant,
+                   uint32_t line, uint32_t column);
 
 symtbl_entry_t *symtbl_lookup(symtbl_t *, str);
 
-void symtbl_add_reference(symtbl_t *, str, size_t line,
-                          size_t column);
+void symtbl_add_reference(symtbl_t *, str, uint32_t line, uint32_t column);
 
 void symtbl_dump(symtbl_t *);
-
-
-/* typedef struct { */
-/*   e_ika_type type_of_number; */
-/*   union { */
-/*     uint8_t uint8; */
-/*     uint16_t uint16; */
-/*     uint32_t uint32; */
-/*     uint64_t uint64; */
-/*     int8_t int8; */
-/*     int16_t int16; */
-/*     int32_t int32; */
-/*     int64_t int64; */
-/*     float flt16; */
-/*     long flt32; */
-/*     long long flt64; */
-/*   }; */
-/* } ika_numeric_value_t; */
-
-/* typedef struct { */
-/*   uint8_t rune[4]; */
-/* } ika_rune_value_t; */
-
-/* typedef struct { */
-/*   uint8_t *raw_bytes; */
-/*   uint64_t length; */
-/* } ika_string_value_t; */
-
-/* typedef struct { */
-/*   e_ika_type type; */
-/*   union { */
-/*     ika_numeric_value_t number; */
-/*     ika_rune_value_t rune; */
-/*     ika_string_value_t string; */
-/*   }; */
-/* } ika_typed_literal_t; */
-
-/* typedef struct { */
-/*   e_ika_type return_type; */
-/*   e_ika_type *parameters; */
-/*   size_t parameter_count; */
-/* } ika_typed_function_definition_t; */
-
-/* typedef struct { */
-/*   e_ika_type type; */
-/*   str *identifier; */
-/* } ika_typed_identifier_t; */
-
-/* typedef struct { */
-/*   e_ika_type type; */
-/*   expr_t *expr; */
-/* } ika_typed_expression_t; */
-
-/* typedef enum { */
-/*   ika_typed_literal, */
-/*   ika_typed_identifier, */
-/*   ika_typed_function_definition, */
-/*   ika_typed_expression, */
-/* } ika_typed_value; */
-
-/* typedef struct { */
-/*   ika_typed_value type; */
-/*   union { */
-/*     ika_typed_literal_t literal; */
-/*     ika_typed_function_definition_t function; */
-/*     ika_typed_expression_t expression; */
-/*     ika_typed_identifier_t identifier; */
-/*   }; */
-/* } symtbl_entry_t; */
